feat: vanilla-theme with sprinkles/colorMode/types
