--- examples/react-basic/src/components/DessertBoxDemo.tsx
@@ -1,5 +1,0 @@
-import { DessertBox } from "../theme/DessertBox";
-
-export const DessertBoxDemo = () => {
-    return <DessertBox color="blue.100">blue.100</DessertBox>;
-};
