--- examples/react-basic/src/components/VanillaThemeDemo.tsx
@@ -1,5 +1,0 @@
-import { DessertBox } from "../theme/DessertBox";
-
-export const VanillaThemeDemo = () => {
-    return <DessertBox color="blue.100">just a blue box</DessertBox>;
-};
