--- examples/react-basic/src/theme/DessertBox.tsx
@@ -1,8 +1,0 @@
-import { createBox } from "@dessert-box/react";
-import type { ComponentProps } from "react";
-
-import { themeSprinkles } from "./sprinkles.css";
-
-export const DessertBox = createBox({ atoms: themeSprinkles });
-
-export type BoxProps = ComponentProps<typeof DessertBox>;
