+++ examples/react-vite-plugin-ssr/src/renderer/_default.page.server.tsx
@@ -3,10 +3,12 @@ import { PageShell } from "./PageShell";
 import { escapeInject, dangerouslySkipEscape } from "vite-plugin-ssr";
 import logoUrl from "./logo.svg";
 import type { PageContextServer } from "./types";
-import { colorMode } from "../theme/color-mode.css";
+import { lightMode, darkMode } from "@box-extractor/vanilla-theme/css";
 
 // See https://vite-plugin-ssr.com/data-fetching
 export const passToClient = ["pageProps", "urlPathname"];
+
+const key = "react-vite-plugin-ssr";
 
 export async function render(pageContext: PageContextServer) {
     const { Page, pageProps } = pageContext;
@@ -27,11 +29,7 @@ export async function render(pageContext: PageContextServer) {
     <html lang="en">
       <head>
         <script>
-        ((d)=>{try{var p=localStorage.getItem('${colorMode.localStorageKey}');if(p==d||(p!='${
-        colorMode.light
-    }'&&matchMedia('(prefers-color-scheme:dark)').matches)) document.documentElement.classList.add(d)}catch(e){}})('${
-        colorMode.dark
-    }')
+        ((d)=>{try{var p=localStorage.getItem('${key}');if(p==d||(p!='${lightMode}'&&matchMedia('(prefers-color-scheme:dark)').matches)) document.documentElement.classList.add(d)}catch(e){}})('${darkMode}')
         </script>
         <meta charset="UTF-8" />
         <link rel="icon" href="${logoUrl}" />
