+++ packages/vanilla-extract/package.json
@@ -65,7 +65,9 @@
             "vite-plugin.ts",
             "esbuild-plugin.ts"
         ],
-        "exports": true
+        "___experimentalFlags_WILL_CHANGE_IN_PATCH": {
+            "exports": true
+        }
     },
     "files": [
         "src",
