+++ packages/vanilla-theme/css/package.json
@@ -1,0 +1,3 @@
+{
+    "main": "../src/css/index.ts"
+}
