+++ packages/vanilla-theme/src/css.ts
@@ -1,0 +1,1 @@
+export * from "./css/index";
