+++ packages/vanilla-theme/src/css/index.ts
@@ -1,0 +1,2 @@
+export * from "./color-mode.css";
+export * from "./sprinkles.css";
