+++ packages/vanilla-theme/src/css/sprinkles.css.ts
@@ -1,6 +1,6 @@
 // import { createSprinkles } from "@vanilla-extract/sprinkles";
 
-import { interactiveProperties, staticProperties } from "./base.css";
+import { interactiveProperties, staticProperties } from "./properties.css";
 import { createBoxSprinkles } from "@box-extractor/vanilla-extract";
 
 export const themeSprinkles = createBoxSprinkles(staticProperties, interactiveProperties);
