+++ packages/vanilla-theme/src/index.ts
@@ -1,0 +1,2 @@
+export { tokens } from "./tokens";
+export * from "./types";
